fix: move storage settings to configs filder
